# backend/dental/openApi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "This is the api for the dental module",
    "title" : "Dental Api",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8098"
  } ],
  "paths" : {
...
    "/children" : {
      "get" : {
        "operationId" : "getChildren",
        "parameters" : [ {
          "in" : "query",
          "name" : "yearFilter",
          "required" : false,
          "schema" : {
            "minimum" : 1900,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "institutionIdFilter",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "groupNameFilter",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "noGroupFilter",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "procedureLabelsFilter",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "excludedProcedureLabelsFilter",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            }
          }
        }, {
          "in" : "query",
          "name" : "sortKey",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/ChildSortKey"
          }
        }, {
          "in" : "query",
          "name" : "sortDirection",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        }, {
          "in" : "query",
          "name" : "pageNumber",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "searchFirstName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "searchLastName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "searchDateOfBirth",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetChildrenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns a list of all dental procedures including child data",
        "tags" : [ "Child" ]
      },
      "post" : {
        "operationId" : "createChild",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateChildRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CreateChildResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Creates a procedure for a child",
        "tags" : [ "Child" ]
      }
    },
    "/children/by-person-id" : {
      "get" : {
        "operationId" : "getChildrenByPerson",
        "parameters" : [ {
          "in" : "query",
          "name" : "personId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetChildrenWithDetailsResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns a list of children including personal data from the central file",
        "tags" : [ "Child" ]
      }
    },
    "/children/children-for-transition/{institutionId}" : {
      "get" : {
        "operationId" : "getChildrenForSchoolYearTransition",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "sortKey",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/ChildForTransitionSortKey"
          }
        }, {
          "in" : "query",
          "name" : "sortDirection",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetChildrenForSchoolYearTransitionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Child" ]
      }
    },
    "/children/daycares-for-transition" : {
      "get" : {
        "operationId" : "getDaycaresForSchoolYearTransition",
        "parameters" : [ {
          "in" : "query",
          "name" : "sortKey",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SchoolYearTransitionSortKey"
          }
        }, {
          "in" : "query",
          "name" : "sortDirection",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        }, {
          "in" : "query",
          "name" : "pageNumber",
          "required" : false,
          "schema" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "in" : "query",
          "name" : "statusFilter",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SchoolYearTransitionStatus"
          }
        }, {
          "in" : "query",
          "name" : "institutionName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetSchoolYearTransitionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Child" ]
      }
    },
    "/children/examination/{examinationId}" : {
      "get" : {
        "operationId" : "getExamination",
        "parameters" : [ {
          "in" : "path",
          "name" : "examinationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Examination"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns an examination based on its identifier",
        "tags" : [ "Child" ]
      },
      "put" : {
        "operationId" : "updateExamination",
        "parameters" : [ {
          "in" : "path",
          "name" : "examinationId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateExaminationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Examination"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Updates an examination",
        "tags" : [ "Child" ]
      }
    },
    "/children/export/{institutionId}" : {
      "get" : {
        "operationId" : "exportChildData",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "groupName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "schoolYear",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Exports child data",
        "tags" : [ "Child" ]
      }
    },
    "/children/fluoridation-consent-bulk" : {
      "put" : {
        "operationId" : "updateFluoridationConsentInBulk",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateFluoridationConsentBulkRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Updates fluoridation consent in bulk",
        "tags" : [ "Child" ]
      }
    },
    "/children/groups-for-transition/{institutionId}" : {
      "get" : {
        "operationId" : "getGroupsForSchoolYearTransition",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetGroupsForSchoolYearTransitionResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "Child" ]
      }
    },
...
    "/children/import/{institutionId}" : {
      "post" : {
        "operationId" : "importXlsx",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "schoolYear",
          "required" : true,
          "schema" : {
            "minimum" : 1900,
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Upload an XLSX file to create multiple children.",
        "tags" : [ "Child" ]
      }
    },
    "/children/institutions/{institutionId}/children" : {
      "get" : {
        "operationId" : "searchChildren",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "searchString",
          "required" : true,
          "schema" : {
            "minLength" : 1,
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SearchChildrenResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Searches and returns children associated with the given institution",
        "tags" : [ "Child" ]
      }
    },
    "/children/institutions/{institutionId}/groups" : {
      "get" : {
        "operationId" : "getInstitutionGroups",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "openGroupsOnly",
          "required" : true,
          "schema" : {
            "type" : "boolean"
          }
        }, {
          "in" : "query",
          "name" : "year",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetInstitutionGroupsResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Returns all created groups of an institution",
        "tags" : [ "Child" ]
      }
    },
    "/children/institutions/{institutionId}/validate" : {
      "get" : {
        "operationId" : "validateOpenChildrenExistInInstitution",
        "parameters" : [ {
          "in" : "path",
          "name" : "institutionId",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "query",
          "name" : "year",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "boolean"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Validates open children exist in institution",
        "tags" : [ "Child" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ChildForTransitionSortKey" : {
        "type" : "string",
        "enum" : [ "ID", "FIRST_NAME", "LAST_NAME", "DATE_OF_BIRTH", "GROUP_NAME" ]
      },
...
      "ChildSortKey" : {
        "type" : "string",
        "enum" : [ "ID", "DATE_OF_BIRTH", "FIRST_NAME", "LAST_NAME", "YEAR", "GROUP_NAME" ]
      },
...
      "CreateChildResponse" : {
        "required" : [ "id" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "Id of the child.",
            "format" : "uuid",
            "example" : "UUID_1"
          }
        }
      },
...
      "GetChildrenResponse" : {
        "required" : [ "elements", "totalNumberOfElements" ],
        "type" : "object",
        "properties" : {
          "elements" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Child"
            }
          },
          "totalNumberOfElements" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "GetChildrenWithDetailsResponse" : {
        "required" : [ "children" ],
        "type" : "object",
        "properties" : {
          "children" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Child"
            }
          }
        }
      },
...
      "SchoolYearTransitionSortKey" : {
        "type" : "string",
        "enum" : [ "ID", "NAME", "COMPLETED_COUNT", "STATUS" ]
      },
...
      "SortDirection" : {
        "type" : "string",
        "enum" : [ "ASC", "DESC" ]
      },
...
    }
  }
}